refactor: split ApplicationReview monolith into focused feature modules (ADS-256)
Extract VisitScheduling, ReferenceChecks, ApplicationTimeline, and
ApplicationDetails into self-contained feature modules, each with a hook,
presentational component, container, and behaviour tests. ApplicationReview.tsx
reduced from 2,145 lines to 298 lines using the container/presentational pattern.
